+++ app/src/main/java/org/sopt/androidassingment/data/BindingConversions.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassingment
+package org.sopt.androidassingment.data
 
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
