+++ app/src/main/java/org/sopt/androidassingment/data/response/ResponseFollowerData.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassingment.data
+package org.sopt.androidassingment.data.response
 
 data class ResponseFollowerData(
     val login : String
