+++ app/src/main/java/org/sopt/androidassingment/data/response/ResponseRepoData.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassingment.data
+package org.sopt.androidassingment.data.response
 
 data class ResponseRepoData(
     val name : String,
