+++ app/src/main/java/org/sopt/androidassingment/data/response/ResponseSignUpData.kt
@@ -1,6 +1,6 @@
-package org.sopt.androidassingment.data
+package org.sopt.androidassingment.data.response
 
-data class ResponseLoginData(
+data class ResponseSignUpData(
     val status: Int,
     val success: Boolean,
     val message: String,
