+++ app/src/main/java/org/sopt/androidassingment/data/server/LoginService.kt
@@ -1,7 +1,7 @@
-package org.sopt.androidassingment.server
+package org.sopt.androidassingment.data.server
 
-import org.sopt.androidassingment.data.RequestLoginData
-import org.sopt.androidassingment.data.ResponseLoginData
+import org.sopt.androidassingment.data.request.RequestLoginData
+import org.sopt.androidassingment.data.response.ResponseLoginData
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.Headers
