+++ app/src/main/java/org/sopt/androidassingment/ui/camera/CameraFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassingment.fragment
+package org.sopt.androidassingment.ui.camera
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,12 +7,14 @@ import android.view.View
 import android.view.ViewGroup
 import org.sopt.androidassingment.R
 
-class HomeFollowerFragment : Fragment() {
+
+class CameraFragment : Fragment() {
+
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home_follower, container, false)
+        return inflater.inflate(R.layout.fragment_camera, container, false)
     }
 }
