+++ app/src/main/java/org/sopt/androidassingment/ui/home/HomeFollowerFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassingment.fragment
+package org.sopt.androidassingment.ui.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,14 +7,12 @@ import android.view.View
 import android.view.ViewGroup
 import org.sopt.androidassingment.R
 
-
-class HomeFollowingFragment : Fragment() {
+class HomeFollowerFragment : Fragment() {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home_following, container, false)
+        return inflater.inflate(R.layout.fragment_home_follower, container, false)
     }
-
 }
