+++ app/src/main/java/org/sopt/androidassingment/ui/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassingment.fragment
+package org.sopt.androidassingment.ui.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,8 +6,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.google.android.material.tabs.TabLayoutMediator
-import org.sopt.androidassingment.adapter.HomeTabViewPagerAdapter
 import org.sopt.androidassingment.databinding.FragmentHomeBinding
+import org.sopt.androidassingment.ui.home.adapter.HomeTabViewPagerAdapter
 
 class HomeFragment : Fragment() {
     private var _binding: FragmentHomeBinding? = null
