+++ app/src/main/res/layout/fragment_home_follower.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/gray2"
-    tools:context=".fragment.HomeFollowerFragment">
+    tools:context=".ui.home.HomeFollowerFragment">
 
 
     <TextView
